+++ hs-opentelemetry-awsxray.cabal
@@ -27,7 +27,9 @@ source-repository head
 
 library
   exposed-modules:
-      OpenTelemetry.AWSXRay
+      OpenTelemetry.AWSXRay.IdGenerator
+      OpenTelemetry.AWSXRay.Propagator
+      OpenTelemetry.AWSXRay.TraceInfo
   other-modules:
       Paths_hs_opentelemetry_awsxray
   hs-source-dirs:
@@ -61,13 +63,17 @@ library
   build-depends:
       base <5
     , bytestring
+    , errors
     , hs-opentelemetry-api
     , hs-opentelemetry-sdk
     , http-types
     , memory
     , random
     , time
+    , unliftio
   default-language: Haskell2010
+  if impl(ghc >= 9.0)
+    ghc-options: -Wno-missing-kind-signatures
   if impl(ghc >= 8.10)
     ghc-options: -Wno-missing-safe-haskell-mode -Wno-prepositive-qualified-module
 
@@ -108,5 +114,7 @@ test-suite spec
       base <5
     , hspec
   default-language: Haskell2010
+  if impl(ghc >= 9.0)
+    ghc-options: -Wno-missing-kind-signatures
   if impl(ghc >= 8.10)
     ghc-options: -Wno-missing-safe-haskell-mode -Wno-prepositive-qualified-module
