+++ package.yaml
@@ -23,6 +23,9 @@ ghc-options:
   - -Wno-safe
 
 when:
+  - condition: "impl(ghc >= 9.0)"
+    ghc-options:
+      - -Wno-missing-kind-signatures
   - condition: "impl(ghc >= 8.10)"
     ghc-options:
       - -Wno-missing-safe-haskell-mode
@@ -61,12 +64,14 @@ library:
   source-dirs: library
   dependencies:
     - bytestring
+    - errors
     - hs-opentelemetry-api
     - hs-opentelemetry-sdk
     - http-types
     - memory
     - random
     - time
+    - unliftio
 
 tests:
   spec:
